requirements.txt
pandas
matplotlib
sqlalchemy
pymysql
python-dotenv
jupyter